[#48] Got annoyed with Vertex class in the graph header file, trying to figure out why the program crashes when trying to insert the vertex during the while loop in shortest path function.
